fixed a bug that makes -infinity sense
so, the practice bot was having serious errors (NI-RIO errors) that
concerned a talon and the gyro. I fixed it by checking back revisions
until it worked, and found out that the source of the problem was that
the invert function on the open talon for the center wheel just broke
everything, so it is not inverted and instead negated before being set.
But wow, did not expect that to cause such a crazy error.